requirements.txt
openai
langchain
faiss-cpu
PyPDF2
tiktoken
python-dotenv
streamlit